Use $doCheck to replace collection thumbnails
Had been using $interval and replacing images, and this was slowing down Primo. New strategy is to hide thumbnails be default, load them in "after" directive, and then unhide default thumbnails in cases where they have not been replaced. Instead of $interval, $doCheck is used to run the process, which seems to be less demanding.